requirements.txt
fastapi==0.104.1
uvicorn==0.20.0
peewee==3.16.3
passlib==1.7.4
pydantic==1.10.13
python-dotenv==1.0.0
psycopg2==2.9.9